backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-db-logger==0.1.12
